refactoring some function to utils
